fix(agent): address streaming audit — token truncation, reasoning sig, diagnostics
Critical fixes from streaming migration audit:

1. CRITICAL: u64→u32 token truncation eliminated
   - Session::add_assistant_message_with_metadata() now accepts ir::Usage
   - execution.rs and streaming.rs pass ir::Usage directly
   - HookInput::post_message() widened to u64 token params

2. HIGH: ReasoningSignature accumulation
   - StreamState accumulates reasoning text + signature
   - do_handle_stream_end writes ContentPart::Reasoning to session
   - Fixes Anthropic extended thinking + tool use validation

3. MEDIUM: Diagnostic warnings added
   - Malformed tool call JSON: warn + log instead of silent fallback
   - ToolCallEnd without Start: warn instead of silent drop
   - PostMessage hook receives FinishReason from Finish chunk

4 files, 1570 lib tests pass, clippy 0 warnings.

+++ src/hooks/traits.rs
@@ -170,8 +170,8 @@ pub enum HookEventData {
     PostMessage {
         model: String,
         stop_reason: Option<String>,
-        input_tokens: u32,
-        output_tokens: u32,
+        input_tokens: u64,
+        output_tokens: u64,
     },
     SessionFork {
         parent_session_id: String,
@@ -406,8 +406,8 @@ impl HookInput {
         session_id: impl Into<String>,
         model: impl Into<String>,
         stop_reason: Option<String>,
-        input_tokens: u32,
-        output_tokens: u32,
+        input_tokens: u64,
+        output_tokens: u64,
     ) -> Self {
         Self::new(
             session_id,
